Another Win64 fix.
TBR=mstarzinger@chromium.org

Review URL: https://codereview.chromium.org/444643002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@22871 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/x64/code-generator-x64.cc
@@ -687,7 +687,7 @@ void CodeGenerator::AssembleArchBoolean(Instruction* instr,
   // last output of the instruction.
   Label check;
   DCHECK_NE(0, instr->OutputCount());
-  Register reg = i.OutputRegister(instr->OutputCount() - 1);
+  Register reg = i.OutputRegister(static_cast<int>(instr->OutputCount() - 1));
   Condition cc = no_condition;
   switch (condition) {
     case kUnorderedEqual:
